`TypeOrGenericTypeArgument` struct moved to separate file.
